+++ pyrad/client_async.py
@@ -241,7 +241,7 @@ class ClientAsync:
 
             acct_connect = self.loop.create_datagram_endpoint(
                 self.protocol_acct,
-                reuse_address=True, reuse_port=True,
+                reuse_port=True,
                 remote_addr=(self.server, self.acct_port),
                 local_addr=bind_addr
             )
@@ -261,7 +261,7 @@ class ClientAsync:
 
             auth_connect = self.loop.create_datagram_endpoint(
                 self.protocol_auth,
-                reuse_address=True, reuse_port=True,
+                reuse_port=True,
                 remote_addr=(self.server, self.auth_port),
                 local_addr=bind_addr
             )
@@ -281,7 +281,7 @@ class ClientAsync:
 
             coa_connect = self.loop.create_datagram_endpoint(
                 self.protocol_coa,
-                reuse_address=True, reuse_port=True,
+                reuse_port=True,
                 remote_addr=(self.server, self.coa_port),
                 local_addr=bind_addr
             )