+++ pyrad/tools.py
@@ -110,11 +110,10 @@ def EncodeAscendBinary(str):
             terms[key] = struct.pack('B', int(value))
 
     trailer = 8 * b'\x00'
-    result = b'%s%s%s\x00%s%s%s%s%s\x00%s%s%s%s\x00\x00%s' % (
-        terms['family'], terms['action'], terms['direction'], terms['src'],
-        terms['dst'], terms['srcl'], terms['dstl'], terms['proto'],
-        terms['sport'], terms['dport'], terms['sportq'], terms['dportq'],
-        trailer)
+
+    result = b''.join((terms['family'], terms['action'], terms['direction'], b'\x00', 
+        terms['src'], terms['dst'], terms['srcl'], terms['dstl'], terms['proto'], b'\x00',
+        terms['sport'], terms['dport'], terms['sportq'], terms['dportq'], b'\x00\x00', trailer))
     return result